pyproject.toml
[project]
name = "weekly-metrics-dashboard"
version = "0.1.0"
description = "Weekly SEO/GEO metrics dashboard for marketing team"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.45.0",
    "pandas>=2.2.0",
    "plotly>=6.0.0",
    "anthropic>=0.52.0",
    "python-dotenv>=1.1.0",
    "google-api-python-client>=2.130.0",
    "google-analytics-data>=0.18.0",
    "google-auth>=2.29.0",
]
